Fix rebase of feature/text-selection-toolbar
* store split up so new stores needed to go to new file: annotation.js
* remove lib/data/stores.js for this branch too
* fix calculation of content height in lib/routes/+page.svelte
